requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
certifi==2024.2.2
click==8.1.7
coverage==7.4.4
ecdsa==0.18.0
exceptiongroup==1.2.0
Faker==24.7.1
fastapi==0.104.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
pyasn1==0.6.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.35
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.23.2
